fix(colab): Share dialog alignment, QR filename, live access-request sync (§22)
- Align Annotation Label / Annotation URL to the same top edge (items-start
  instead of items-end, which let a select/input height mismatch push one
  label lower than the other).
- Name QR downloads after the dataset and label (e.g. hpa-demo-dataset-cells.png)
  instead of a generic qr-code.png.
- Poll the broker for pending access requests every 20s while the Share
  dialog is open, so a new request appears without reopening the dialog.
- Move the pending access requests list into the Sharing box, alongside
  owner/managers/annotators/public, instead of its own standalone section.
- Move the Apply button into the dialog's footer next to Close, as the
  final action row.
- Annotate page: poll the broker every 12s while the "Access needed" dialog
  is shown, so the dialog clears itself and the page unblocks as soon as an
  owner grants access, with no manual reload.
